[SEAM] JournalStreamDesc.max_scopes: a stream declares its own RAM bound, and the register bypass dies
Recommendation 1 line 5 of gate/CORE-HOMES-design.md.

THE DEFECT, in the code's own words: "The `JournalStreamDesc` ABI POD carries no
cap, and this crate must not touch the hot ABI to add one, so an in-core plane seam
user reaches this directly instead of the vtable's `journal_register`"
(plane_host/journal.rs, `journal_register_capped`). One caller — the MCP call log —
and it crossed the seam sideways because a POD was missing a field. A plane's tuning
constant that the ABI cannot express is an ABI hole, not a plane's problem.

THE FACE: `JournalStreamDesc._reserved: u32` becomes `max_scopes: u32` — the LRU
bound on the host's RAM position cache for that stream, `0` = the host default
(unbounded). This is WEAKER than the append the design note priced: `_reserved` was
declared padding before the borrowed range, so every field offset and `__size` (32)
are UNCHANGED, and a peer built against minor-20 wrote `0` into that slot, which
decodes to exactly the host default it already received. `ABI_MINOR` 20->21 records
it; `POD_VERSION` is unmoved for the same reason; the layout golden reseeds ONE line
(`JournalStreamDesc._reserved=12` -> `.max_scopes=12`, same offset).

THE CONSEQUENCE: `journal_register` reads `d.max_scopes` as data, `register_stream`
loses its `cap` parameter, and `journal_register_capped` is DELETED. `calllog` now
registers through the vtable slot like every other stream, declaring
`max_scopes: MAX_TRACKED_PRINCIPALS` (16,384) in its own descriptor — the constant
stays where it is for now, but it is now a DECLARED field of the registration rather
than a number core reaches around the ABI to apply, which is what line 8 needs to
carry the file to busbar-plane-mcp without a second entry point following it.
`plane::auditlog` and the three test descriptors declare `max_scopes: 0` — the
default they already had.

Proven: `cargo test -p busbar-plugin` (layout golden green after the one-line
reseed), `cargo test -p busbar-core --lib` 1872 pass / 0 fail — including
`frozen_chains_boot_verify_through_the_durable_seam` and the two calllog LRU-bound
tests that pin the 16,384 cap end-to-end through the new descriptor field.

One busbar-core kind-isolation cell is RE-PINNED DOWN — `journal_register_capped`
and its within-core call site took three plane words with them, and the matrix row
refuses stale slack in either direction: busbar-core x plane 3026->3023. Down-only;
no declaration owed. No ceiling RISES on this line: busbar-plugin carries no §1.1
ceiling (it is not in `kernel_crate` / `caps_crate` / `contract_crate` /
`unit_crate_glob` / `verbs_crate`), and no unit, kernel, caps or contract crate is
touched, so `unit_total_ceiling` 14933 and `union_ceiling` 23825 are UNMOVED.
`cargo xtask gate kind-isolation` 11/11 green at this commit.

+++ crates/busbar-core/src/tests/calllog_tests.rs
@@ -1317,7 +1317,7 @@ fn the_chain_position_map_stays_bounded_across_many_distinct_principals() {
     }
     assert_eq!(
         log.len(),
-        MAX_TRACKED_PRINCIPALS,
+        MAX_TRACKED_PRINCIPALS as usize,
         "the map is capped at MAX_TRACKED_PRINCIPALS however many distinct principals arrive, \
          rather than growing one entry per principal forever"
     );
@@ -1360,7 +1360,7 @@ fn an_evicted_principal_resumes_from_the_store_instead_of_forking_its_chain() {
     );
     assert_eq!(
         log.len(),
-        MAX_TRACKED_PRINCIPALS,
+        MAX_TRACKED_PRINCIPALS as usize,
         "still capped after the flood"
     );
